edge/raspiImageGenerator/pyproject.toml
[tool.poetry]
name = "w4b-raspi-image-generator"
version = "0.1.0"
description = "Raspberry Pi Image Generator for the W4B Hive Monitoring System"
authors = ["Your Name <your.email@example.com>"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "w4b_image_generator"}]
repository = "https://github.com/itsatony/w4b_v3"

[tool.poetry.dependencies]
python = "^3.9"
asyncio = "^3.4.3"
aiohttp = "^3.8.1"
PyYAML = "^6.0"
click = "^8.1.3"
jinja2 = "^3.1.2"
cryptography = "^38.0.1"
paramiko = "^2.11.0"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.20.0"
pytest-cov = "^4.0.0"
black = "^22.10.0"
mypy = "^0.982"
flake8 = "^5.0.4"
bandit = "^1.7.4"
isort = "^5.10.1"
pre-commit = "^2.20.0"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.3.0"
sphinx-rtd-theme = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.poetry.scripts]
w4b-image-generator = "w4b_image_generator.image_generator:main_cli"
